Share Windows GlobalMemory/VirtualMemory logic across JNA and FFM
Windows was the last platform whose memory twins still carried full logic in
both backends; every other platform already has a base + thin native-only
twins. Bring Windows into line.

WindowsGlobalMemory (existing base, previously only the memory-type tables)
now holds the memoized available/total/page-size suppliers, the virtual-memory
supplier, and the WMI physical-memory assembly loop, with protected abstract
hooks for the native queries (readPerfInfo, createVirtualMemory,
isWindows10OrGreater, queryPhysicalMemory[Win8]).

WindowsVirtualMemory is new (VirtualMemory had no shared base): it holds the
parent-global reference, memoized suppliers, and the six page-size-scaled
getters, with hooks querySwapUsed / querySwapTotalVirtMaxVirtUsed /
queryPageSwaps.

The JNA and FFM twins shrink to just those native hooks (GlobalMemory
119/128 -> 67/79 lines, VirtualMemory 106/111 -> 65/71). No behavior change;
the hoisted logic is identical to what both twins previously duplicated.
