+++ oshi-common/src/main/java/oshi/hardware/common/platform/windows/WindowsGlobalMemory.java
@@ -4,11 +4,27 @@
  */
 package oshi.hardware.common.platform.windows;
 
+import static oshi.util.Memoizer.defaultExpiration;
+import static oshi.util.Memoizer.memoize;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.function.Supplier;
+
 import oshi.annotation.concurrent.ThreadSafe;
+import oshi.driver.common.windows.wmi.Win32PhysicalMemory.PhysicalMemoryProperty;
+import oshi.driver.common.windows.wmi.Win32PhysicalMemory.PhysicalMemoryPropertyWin8;
+import oshi.driver.common.windows.wmi.WmiResult;
+import oshi.driver.common.windows.wmi.WmiUtil;
+import oshi.hardware.PhysicalMemory;
+import oshi.hardware.VirtualMemory;
 import oshi.hardware.common.AbstractGlobalMemory;
+import oshi.util.tuples.Triplet;
 
 /**
- * Common Windows global memory logic shared between JNA and FFM implementations.
+ * Common Windows global memory logic shared between JNA and FFM implementations. Subclasses supply the native queries
+ * (performance info, WMI physical-memory banks, virtual memory) via the {@code protected abstract} hooks; the
+ * memoization and result assembly live here.
  */
 @ThreadSafe
 public abstract class WindowsGlobalMemory extends AbstractGlobalMemory {
@@ -16,6 +32,98 @@ public abstract class WindowsGlobalMemory extends AbstractGlobalMemory {
     /** Default constructor. */
     protected WindowsGlobalMemory() {
     }
+
+    private final Supplier<Triplet<Long, Long, Long>> availTotalSize = memoize(this::readPerfInfo, defaultExpiration());
+
+    private final Supplier<VirtualMemory> vm = memoize(this::createVirtualMemory);
+
+    @Override
+    public long getAvailable() {
+        return availTotalSize.get().getA();
+    }
+
+    @Override
+    public long getTotal() {
+        return availTotalSize.get().getB();
+    }
+
+    @Override
+    public long getPageSize() {
+        return availTotalSize.get().getC();
+    }
+
+    @Override
+    public VirtualMemory getVirtualMemory() {
+        return vm.get();
+    }
+
+    @Override
+    public List<PhysicalMemory> getPhysicalMemory() {
+        List<PhysicalMemory> physicalMemoryList = new ArrayList<>();
+        if (isWindows10OrGreater()) {
+            WmiResult<PhysicalMemoryProperty> bankMap = queryPhysicalMemory();
+            for (int index = 0; index < bankMap.getResultCount(); index++) {
+                String bankLabel = WmiUtil.getString(bankMap, PhysicalMemoryProperty.BANKLABEL, index);
+                long capacity = WmiUtil.getUint64(bankMap, PhysicalMemoryProperty.CAPACITY, index);
+                long speed = WmiUtil.getUint32(bankMap, PhysicalMemoryProperty.SPEED, index) * 1_000_000L;
+                String manufacturer = WmiUtil.getString(bankMap, PhysicalMemoryProperty.MANUFACTURER, index);
+                String memoryType = smBiosMemoryType(
+                        WmiUtil.getUint32(bankMap, PhysicalMemoryProperty.SMBIOSMEMORYTYPE, index));
+                String partNumber = WmiUtil.getString(bankMap, PhysicalMemoryProperty.PARTNUMBER, index);
+                String serialNumber = WmiUtil.getString(bankMap, PhysicalMemoryProperty.SERIALNUMBER, index);
+                physicalMemoryList.add(new PhysicalMemory(bankLabel, capacity, speed, manufacturer, memoryType,
+                        partNumber, serialNumber));
+            }
+        } else {
+            WmiResult<PhysicalMemoryPropertyWin8> bankMap = queryPhysicalMemoryWin8();
+            for (int index = 0; index < bankMap.getResultCount(); index++) {
+                String bankLabel = WmiUtil.getString(bankMap, PhysicalMemoryPropertyWin8.BANKLABEL, index);
+                long capacity = WmiUtil.getUint64(bankMap, PhysicalMemoryPropertyWin8.CAPACITY, index);
+                long speed = WmiUtil.getUint32(bankMap, PhysicalMemoryPropertyWin8.SPEED, index) * 1_000_000L;
+                String manufacturer = WmiUtil.getString(bankMap, PhysicalMemoryPropertyWin8.MANUFACTURER, index);
+                String memoryType = memoryType(
+                        WmiUtil.getUint16(bankMap, PhysicalMemoryPropertyWin8.MEMORYTYPE, index));
+                String partNumber = WmiUtil.getString(bankMap, PhysicalMemoryPropertyWin8.PARTNUMBER, index);
+                String serialNumber = WmiUtil.getString(bankMap, PhysicalMemoryPropertyWin8.SERIALNUMBER, index);
+                physicalMemoryList.add(new PhysicalMemory(bankLabel, capacity, speed, manufacturer, memoryType,
+                        partNumber, serialNumber));
+            }
+        }
+        return physicalMemoryList;
+    }
+
+    /**
+     * Queries performance info for the available, total, and page-size values.
+     *
+     * @return a triplet of (available bytes, total bytes, page size)
+     */
+    protected abstract Triplet<Long, Long, Long> readPerfInfo();
+
+    /**
+     * Creates the backend-specific virtual memory instance.
+     *
+     * @return the virtual memory
+     */
+    protected abstract VirtualMemory createVirtualMemory();
+
+    /**
+     * @return {@code true} if running on Windows 10 or greater
+     */
+    protected abstract boolean isWindows10OrGreater();
+
+    /**
+     * Queries the Win32_PhysicalMemory banks (Windows 10+ schema).
+     *
+     * @return the WMI result
+     */
+    protected abstract WmiResult<PhysicalMemoryProperty> queryPhysicalMemory();
+
+    /**
+     * Queries the Win32_PhysicalMemory banks (pre-Windows 10 schema).
+     *
+     * @return the WMI result
+     */
+    protected abstract WmiResult<PhysicalMemoryPropertyWin8> queryPhysicalMemoryWin8();
 
     /**
      * memoryType.
