+++ oshi-core-ffm/src/main/java/oshi/hardware/platform/windows/WindowsGlobalMemoryFFM.java
@@ -6,15 +6,10 @@ package oshi.hardware.platform.windows;
 
 import static org.slf4j.event.Level.ERROR;
 import static oshi.ffm.ForeignFunctions.callInArenaOrDefault;
-import static oshi.util.Memoizer.defaultExpiration;
-import static oshi.util.Memoizer.memoize;
 
 import java.lang.foreign.MemoryLayout;
 import java.lang.foreign.MemorySegment;
 import java.lang.foreign.ValueLayout;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.function.Supplier;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -23,12 +18,10 @@ import oshi.annotation.concurrent.ThreadSafe;
 import oshi.driver.common.windows.wmi.Win32PhysicalMemory.PhysicalMemoryProperty;
 import oshi.driver.common.windows.wmi.Win32PhysicalMemory.PhysicalMemoryPropertyWin8;
 import oshi.driver.common.windows.wmi.WmiResult;
-import oshi.driver.common.windows.wmi.WmiUtil;
 import oshi.driver.windows.wmi.Win32PhysicalMemoryFFM;
 import oshi.ffm.platform.windows.Kernel32FFM;
 import oshi.ffm.platform.windows.PsapiFFM;
 import oshi.ffm.platform.windows.VersionHelpersFFM;
-import oshi.hardware.PhysicalMemory;
 import oshi.hardware.VirtualMemory;
 import oshi.hardware.common.platform.windows.WindowsGlobalMemory;
 import oshi.util.tuples.Triplet;
@@ -43,70 +36,8 @@ final class WindowsGlobalMemoryFFM extends WindowsGlobalMemory {
 
     private static final boolean IS_WINDOWS10_OR_GREATER = VersionHelpersFFM.IsWindows10OrGreater();
 
-    private final Supplier<Triplet<Long, Long, Long>> availTotalSize = memoize(WindowsGlobalMemoryFFM::readPerfInfo,
-            defaultExpiration());
-
-    private final Supplier<VirtualMemory> vm = memoize(this::createVirtualMemory);
-
     @Override
-    public long getAvailable() {
-        return availTotalSize.get().getA();
-    }
-
-    @Override
-    public long getTotal() {
-        return availTotalSize.get().getB();
-    }
-
-    @Override
-    public long getPageSize() {
-        return availTotalSize.get().getC();
-    }
-
-    @Override
-    public VirtualMemory getVirtualMemory() {
-        return vm.get();
-    }
-
-    private VirtualMemory createVirtualMemory() {
-        return new WindowsVirtualMemoryFFM(this);
-    }
-
-    @Override
-    public List<PhysicalMemory> getPhysicalMemory() {
-        List<PhysicalMemory> physicalMemoryList = new ArrayList<>();
-        if (IS_WINDOWS10_OR_GREATER) {
-            WmiResult<PhysicalMemoryProperty> bankMap = Win32PhysicalMemoryFFM.queryPhysicalMemory();
-            for (int index = 0; index < bankMap.getResultCount(); index++) {
-                String bankLabel = WmiUtil.getString(bankMap, PhysicalMemoryProperty.BANKLABEL, index);
-                long capacity = WmiUtil.getUint64(bankMap, PhysicalMemoryProperty.CAPACITY, index);
-                long speed = WmiUtil.getUint32(bankMap, PhysicalMemoryProperty.SPEED, index) * 1_000_000L;
-                String manufacturer = WmiUtil.getString(bankMap, PhysicalMemoryProperty.MANUFACTURER, index);
-                String memType = smBiosMemoryType(
-                        WmiUtil.getUint32(bankMap, PhysicalMemoryProperty.SMBIOSMEMORYTYPE, index));
-                String partNumber = WmiUtil.getString(bankMap, PhysicalMemoryProperty.PARTNUMBER, index);
-                String serialNumber = WmiUtil.getString(bankMap, PhysicalMemoryProperty.SERIALNUMBER, index);
-                physicalMemoryList.add(new PhysicalMemory(bankLabel, capacity, speed, manufacturer, memType, partNumber,
-                        serialNumber));
-            }
-        } else {
-            WmiResult<PhysicalMemoryPropertyWin8> bankMap = Win32PhysicalMemoryFFM.queryPhysicalMemoryWin8();
-            for (int index = 0; index < bankMap.getResultCount(); index++) {
-                String bankLabel = WmiUtil.getString(bankMap, PhysicalMemoryPropertyWin8.BANKLABEL, index);
-                long capacity = WmiUtil.getUint64(bankMap, PhysicalMemoryPropertyWin8.CAPACITY, index);
-                long speed = WmiUtil.getUint32(bankMap, PhysicalMemoryPropertyWin8.SPEED, index) * 1_000_000L;
-                String manufacturer = WmiUtil.getString(bankMap, PhysicalMemoryPropertyWin8.MANUFACTURER, index);
-                String memType = memoryType(WmiUtil.getUint16(bankMap, PhysicalMemoryPropertyWin8.MEMORYTYPE, index));
-                String partNumber = WmiUtil.getString(bankMap, PhysicalMemoryPropertyWin8.PARTNUMBER, index);
-                String serialNumber = WmiUtil.getString(bankMap, PhysicalMemoryPropertyWin8.SERIALNUMBER, index);
-                physicalMemoryList.add(new PhysicalMemory(bankLabel, capacity, speed, manufacturer, memType, partNumber,
-                        serialNumber));
-            }
-        }
-        return physicalMemoryList;
-    }
-
-    private static Triplet<Long, Long, Long> readPerfInfo() {
+    protected Triplet<Long, Long, Long> readPerfInfo() {
         return callInArenaOrDefault(arena -> {
             MemorySegment perfInfo = arena.allocate(PsapiFFM.PERFORMANCE_INFORMATION_LAYOUT);
             int size = (int) PsapiFFM.PERFORMANCE_INFORMATION_LAYOUT.byteSize();
@@ -124,5 +55,25 @@ final class WindowsGlobalMemoryFFM extends WindowsGlobalMemory {
             long memTotal = pageSize * physTotal;
             return new Triplet<>(memAvailable, memTotal, pageSize);
         }, LOG, ERROR, "Failed to get Performance Info", new Triplet<>(0L, 0L, 4096L));
+    }
+
+    @Override
+    protected VirtualMemory createVirtualMemory() {
+        return new WindowsVirtualMemoryFFM(this);
+    }
+
+    @Override
+    protected boolean isWindows10OrGreater() {
+        return IS_WINDOWS10_OR_GREATER;
+    }
+
+    @Override
+    protected WmiResult<PhysicalMemoryProperty> queryPhysicalMemory() {
+        return Win32PhysicalMemoryFFM.queryPhysicalMemory();
+    }
+
+    @Override
+    protected WmiResult<PhysicalMemoryPropertyWin8> queryPhysicalMemoryWin8() {
+        return Win32PhysicalMemoryFFM.queryPhysicalMemoryWin8();
     }
 }
